backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
python-dotenv==1.0.0
pandas>=2.2.0
numpy>=1.26.0
scikit-learn>=1.4.0
langchain>=0.1.0
langchain-openai>=0.0.5
openai>=1.0.0
requests>=2.31.0
Werkzeug>=3.0.0
ta>=0.10.2
joblib>=1.3.0